Fix unlock for Redlock adapter
When lock is deserialized, it is hash containing strings as keys.
Though, Redlock expects lock to have Symbols as keys.
Hence, the job is completed without releasing a lock.

This might be valid for other adapters. (I have not checked it)
If so, the problem must be fixed on `desecialize` step.

Did not find how to set up test environment, so don't have specs,
unfortunately.